error verbose output

+++ Get-DellDrivers.ps1
@@ -18,9 +18,16 @@ foreach ($uri in $url_list)
     # [System.String]$cabfile = ((Invoke-WebRequest -Uri "${uri}" -UserAgent $userAgent -UseBasicParsing).Links | Where-Object -FilterScript {$_.href -match '^http.*(downloads|dl).dell.com/.*-win10-.*\.CAB$'} | Select-Object -First 1 | Select-Object -ExpandProperty outerHTML) -replace '.*(http.*\.CAB).*','$1'
     [System.String]$cabfile = ((Invoke-WebRequest -Uri "${uri}" -UserAgent $userAgent -UseBasicParsing).Links | Where-Object -FilterScript {$_.href -match '^http.*-win10-.*\.CAB'} | Select-Object -First 1 | Select-Object -ExpandProperty outerHTML) -replace '.*(http.*\.CAB).*','$1'
     [System.String]$outFile = "${PSScriptRoot}\Dell\Latitude\win10\${directory}\$(Split-Path -Path $cabfile.Replace('%20',' ') -Leaf)"
-    Invoke-WebRequest -Uri $cabfile -UseBasicParsing -UserAgent $userAgent -ContentType 'application/zip' -OutFile $outFile
+    try {
+        Invoke-WebRequest -Uri $cabfile -UseBasicParsing -UserAgent $userAgent -ContentType 'application/zip' -OutFile $outFile
+    } catch {
+        Write-Output "Unable to download file"
+    }
 
     <# VERIFY DOWNLOAD #>
-    Test-Path -Path $outFile
-    Remove-Item -Path $outFile -Confirm:$false -Force
+    if (Test-Path -Path $outFile)
+    {
+        Remove-Item -Path $outFile -Confirm:$false -Force
+    }
+    
 }
